requirements.txt
requests
python-dotenv
auto-py-to-exe